댓글 삭제 및 추가 로직 구현

+++ src/main/java/com/community/board/controller/ReplyController.java
@@ -14,9 +14,12 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
 import java.io.IOException;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Optional;
 
@@ -31,7 +34,12 @@ public class ReplyController {
 
     private final ReplyService replyService;
 
-    // 좋아요 관련 내용
+    private String updatePath(String path) {
+        return URLEncoder.encode(path, StandardCharsets.UTF_8);
+    }
+
+
+    // 댓글 관련 내용
     @ResponseBody
     @RequestMapping(value = "/board/detail/reply")
     public int addLikeLink(@RequestParam(value = "r_board_id") Long r_board_id,
@@ -59,28 +67,36 @@ public class ReplyController {
         return reply_size;
     }
 
+    @ResponseBody
+    @RequestMapping(value = "/board/detail/reply/update")
+    public int boardReplyUpdate(@RequestParam(value = "reply_update_rid") Long reply_update_rid,
+                                @RequestParam(value = "reply_update_content") String reply_update_content) throws IOException{
+        log.info("rid : " + reply_update_rid);
+        log.info("content : " + reply_update_content);
+        replyService.updateReply(reply_update_rid, reply_update_content);
 
-    /* 댓글 작성 관련 로직 */
-    /*@PostMapping("/board/detail/{boardId}/reply")
-    public String boardReply(@PathVariable Long boardId, @Valid ReplyForm replyForm, @CurrentUser Account account) {
-        Board currentBoard = boardRepository.findByBid(boardId);
-        replyService.saveReply(replyForm, account, currentBoard);
-        return "redirect:/board/detail/{boardId}";
-    }*/
-
-    @PostMapping("/board/detail/{boardId}/reply/update/{rid}")
-    public String boardReplyUpdate(@PathVariable Long boardId,
-                                   @PathVariable Long rid, @Valid ReplyForm replyForm) {
-        replyService.updateReply(rid, replyForm);
-        return "redirect:/board/detail/{boardId}";
+        Reply reply = replyRepository.findByRid(reply_update_rid);
+        Board board = boardRepository.findByBid(reply.getBoard().getBid());
+        List<Reply> replies = replyRepository.findAllByBoard(board);
+        int reply_size = replies.size();
+        return reply_size;
     }
 
-    @PostMapping("/board/detail/{boardId}/reply/delete/{rid}")
-    public String boardReplyDelete(@PathVariable Long boardId,
-                                   @PathVariable Long rid) {
+    @GetMapping("/board/detail/reply/delete/{rid}")
+    public String boardReplyDelete(@PathVariable Long rid,
+                                   RedirectAttributes redirectAttributes) {
+        String r_del_error_message = null;
         Reply findReply = replyRepository.findByRid(rid);
+        Board board = boardRepository.findByBid(findReply.getBoard().getBid());
+        String path = String.valueOf(board.getBid());
+
+        if (findReply.getIsReported() || findReply.getReportCount() > 0) {
+            r_del_error_message = "신고된 댓글은 삭제할 수 없습니다.";
+            return r_del_error_message;
+        }
+        redirectAttributes.addFlashAttribute("r_del_complete_message", "댓글이 삭제되었습니다.");
         replyRepository.delete(findReply);
-        return "redirect:/board/detail/{boardId}";
+        return "redirect:/board/detail/" + updatePath(path);
     }
 
 }
